+++ src/coreComponents/physicsSolvers/PhysicsSolverBase.hpp
@@ -1133,8 +1133,8 @@ void PhysicsSolverBase::setConstitutiveName( ElementSubRegionBase & subRegion, s
 
   string & constitutiveName = subRegion.getReference< string >( wrapperName );
   constitutiveName = getConstitutiveName< CONSTITUTIVE_TYPE >( subRegion );
-  GEOS_ERROR_IF( constitutiveName.empty(), GEOS_FMT( "{}: constitutive model not found on subregion {}",
-                                                     getDataContext(), subRegion.getName() ) );
+  GEOS_ERROR_IF( constitutiveName.empty(), GEOS_FMT( "{}: {} constitutive model not found on subregion {}",
+                                                     getDataContext(), typeid(CONSTITUTIVE_TYPE).name(), subRegion.getName() ) );
 }
 
 } // namespace geos
