+++ src/coreComponents/physicsSolvers/fluidFlow/StencilDataCollection.hpp
@@ -21,10 +21,13 @@
 #define SRC_CORECOMPONENTS_PHYSICSSOLVERS_FLUIDFLOW_CELLTOCELLDATACOLLECTOR_HPP_
 
 #include "events/tasks/TaskBase.hpp"
-#include "physicsSolvers/fluidFlow/FlowSolverBase.hpp"
-
+#include "mesh/ElementRegionManager.hpp"
 namespace geos
 {
+
+class FlowSolverBase;
+class MeshLevel;
+class FluxApproximationBase;
 
 /**
  * @class StencilDataCollection
