+++ src/coreComponents/physicsSolvers/surfaceGeneration/EmbeddedSurfaceGenerator.hpp
@@ -28,18 +28,7 @@
 namespace geos
 {
 
-struct NewObjectLists
-{
-  std::set< localIndex > newNodes;
-  std::set< localIndex > newEdges;
-  map< std::pair< localIndex, localIndex >, std::set< localIndex > > newElements;
-
-  void insert( NewObjectLists const & lists );
-};
-
-
 class SpatialPartition;
-
 class NodeManager;
 class FaceManager;
 class ElementRegionManager;
