+++ src/coreComponents/physicsSolvers/surfaceGeneration/EmbeddedSurfacesParallelSynchronization.cpp
@@ -23,14 +23,15 @@
 #include "common/TimingMacros.hpp"
 #include "mesh/ElementRegionManager.hpp"
 #include "mesh/MeshFields.hpp"
+#include "mesh/MeshLevel.hpp"
 #include "mesh/mpiCommunications/CommunicationTools.hpp"
 #include "mesh/mpiCommunications/MPI_iCommData.hpp"
-
+#include "mesh/mpiCommunications/NeighborCommunicator.hpp"
 
 namespace geos
 {
 
-namespace  embeddedSurfacesParallelSynchronization
+namespace embeddedSurfacesParallelSynchronization
 {
 
 using namespace dataRepository;
@@ -593,11 +594,11 @@ void synchronizeFracturedElements( MeshLevel & mesh,
 
 using namespace parallelSynchronizationHelpers;
 
-void sychronizeTopology( MeshLevel & mesh,
-                         stdVector< NeighborCommunicator > & neighbors,
-                         NewObjectLists & newObjects,
-                         int const mpiCommOrder,
-                         string const fractureRegionName )
+void synchronizeTopology( MeshLevel & mesh,
+                          stdVector< NeighborCommunicator > & neighbors,
+                          NewObjectLists & newObjects,
+                          int const mpiCommOrder,
+                          string const fractureRegionName )
 {
 
   // Synchronize nodes
