+++ src/coreComponents/physicsSolvers/surfaceGeneration/EmbeddedSurfacesParallelSynchronization.hpp
@@ -21,22 +21,47 @@
 #ifndef GEOS_PHYSICSSOLVERS_SURFACEGENERATION_EMBEDDEDSURFACESPARALLELSYNCHRONIZATION_HPP_
 #define GEOS_PHYSICSSOLVERS_SURFACEGENERATION_EMBEDDEDSURFACESPARALLELSYNCHRONIZATION_HPP_
 
-#include "physicsSolvers/surfaceGeneration/EmbeddedSurfaceGenerator.hpp"
+#include "common/DataTypes.hpp"
+#include "common/StdContainerWrappers.hpp"
 
 namespace geos
 {
 class MeshLevel;
 class NeighborCommunicator;
-struct ModifiedObjectLists;
+
+struct NewObjectLists
+{
+  std::set< localIndex > newNodes;
+  std::set< localIndex > newEdges;
+  map< std::pair< localIndex, localIndex >, std::set< localIndex > > newElements;
+
+  void insert( NewObjectLists const & newObjects )
+  {
+    newNodes.insert( newObjects.newNodes.begin(),
+                     newObjects.newNodes.end() );
+
+    newEdges.insert( newObjects.newEdges.begin(),
+                     newObjects.newEdges.end() );
+
+    for( auto & iter : newObjects.newElements )
+    {
+      std::pair< localIndex, localIndex > const & key = iter.first;
+      std::set< localIndex > const & values = iter.second;
+      newElements[key].insert( values.begin(), values.end() );
+    }
+
+  }
+
+};
 
 namespace embeddedSurfacesParallelSynchronization
 {
 
-void sychronizeTopology( MeshLevel & mesh,
-                         stdVector< NeighborCommunicator > & neighbors,
-                         NewObjectLists & newObjects,
-                         int const mpiCommOrder,
-                         string const fractureRegionName );
+void synchronizeTopology( MeshLevel & mesh,
+                          stdVector< NeighborCommunicator > & neighbors,
+                          NewObjectLists & newObjects,
+                          int const mpiCommOrder,
+                          string const fractureRegionName );
 }
 
 } /* namespace geos */
