+++ src/coreComponents/physicsSolvers/surfaceGeneration/ParallelTopologyChange.hpp
@@ -20,7 +20,9 @@
 #ifndef GEOS_PHYSICSSOLVERS_SURFACEGENERATION_PARALLELTOPOLOGYCHANGE_HPP_
 #define GEOS_PHYSICSSOLVERS_SURFACEGENERATION_PARALLELTOPOLOGYCHANGE_HPP_
 
-#include "physicsSolvers/surfaceGeneration/SurfaceGenerator.hpp"
+//#include "physicsSolvers/surfaceGeneration/SurfaceGenerator.hpp"
+#include "common/DataTypes.hpp"
+#include "mesh/ElementRegionManager.hpp"
 
 #define PARALLEL_TOPOLOGY_CHANGE_METHOD 1
 
@@ -28,7 +30,69 @@ namespace geos
 {
 class MeshLevel;
 class NeighborCommunicator;
-struct ModifiedObjectLists;
+
+struct ModifiedObjectLists
+{
+  std::set< localIndex > newNodes;
+  std::set< localIndex > newEdges;
+  std::set< localIndex > newFaces;
+  std::set< localIndex > modifiedNodes;
+  std::set< localIndex > modifiedEdges;
+  std::set< localIndex > modifiedFaces;
+  map< std::pair< localIndex, localIndex >, std::set< localIndex > > newElements;
+  map< std::pair< localIndex, localIndex >, std::set< localIndex > > modifiedElements;
+
+  void clearNewFromModified()
+  {
+    for( localIndex const a : newNodes )
+    {
+      modifiedNodes.erase( a );
+    }
+
+    for( localIndex const a : newEdges )
+    {
+      modifiedEdges.erase( a );
+    }
+
+    for( localIndex const a : newFaces )
+    {
+      modifiedFaces.erase( a );
+    }
+  }
+
+  void insert( ModifiedObjectLists const & modifiedObjects )
+  {
+    newNodes.insert( modifiedObjects.newNodes.begin(),
+                     modifiedObjects.newNodes.end() );
+    modifiedNodes.insert( modifiedObjects.modifiedNodes.begin(),
+                          modifiedObjects.modifiedNodes.end() );
+
+    newEdges.insert( modifiedObjects.newEdges.begin(),
+                     modifiedObjects.newEdges.end() );
+    modifiedEdges.insert( modifiedObjects.modifiedEdges.begin(),
+                          modifiedObjects.modifiedEdges.end() );
+
+    newFaces.insert( modifiedObjects.newFaces.begin(),
+                     modifiedObjects.newFaces.end() );
+    modifiedFaces.insert( modifiedObjects.modifiedFaces.begin(),
+                          modifiedObjects.modifiedFaces.end() );
+
+    for( auto & iter : modifiedObjects.newElements )
+    {
+      std::pair< localIndex, localIndex > const & key = iter.first;
+      std::set< localIndex > const & values = iter.second;
+      newElements[key].insert( values.begin(), values.end() );
+    }
+
+    for( auto & iter : modifiedObjects.modifiedElements )
+    {
+      std::pair< localIndex, localIndex > const & key = iter.first;
+      std::set< localIndex > const & values = iter.second;
+      modifiedElements[key].insert( values.begin(), values.end() );
+    }
+
+  }
+};
 
 namespace parallelTopologyChange
 {
